機器人/requirements.txt
Flask==2.1.2
line-bot-sdk==1.16.0
gunicorn==20.0.4
